Refactor: 타입 선택에 따른 filteredData를 생성하는 useFilteredPokemonData 훅과 스크롤에 따른 visibleData를 생성하는 useVisibleDataByScroll 훅 분리 적용
